block more revisions ...
TO ALL DEVELOPERS WITH COMMIT ACCESS:
Please remember to *always* do an 'svnmerge block -r <rev#>' on the trunk
every single time you make a commit to the 1.2 branch which will not be pulled
into the trunk using svnmerge.  This is important for us to be able to take
full advatage of this tool.