- Add a parkinglot setting to ast_channel - Enabling this for users and peers in sip.conf and chan_sip Todo: - use this in res_features :-)
This is to solve the issue when the callee transfers a call and wants
the call to be parked in her parkinglot, not in the callers parkinglot.
Since we now have multiple parkinglots in one call with two channels,
we have to handle different parkinglots for caller and callee. I can no
longer avoid changing ast_channel, sorry.