Merged revisions 36912,36941 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r36912 | oej | 2006-07-04 08:54:55 -0600 (Tue, 04 Jul 2006) | 2 lines

Specify digest algorithm for picky clients

........
r36941 | crichter | 2006-07-04 11:04:21 -0600 (Tue, 04 Jul 2006) | 1 line

removed the reloading of the ntdebug stuff, since this is a bit buggy, we only set it when loading chan_misdn for now. fixed a litle state problem when receiving RELEASE_COMPLETE. also we may only play tones to a NT when the extension does not match and such cases.
........