big set of changes to hold and music on hold stuff ... all untested and i haven't looked back through my changes very much yet
- deprecate musicclass option in channel drivers
- add mohinterpret and mohsuggest options to various channel drivers
- mohinterpret is the moh class to use when put on hold
- mohsuggest is the moh class to suggest to the peer channel being put on hold
- make the handling of HOLD and UNHOLD frames more consistent, meaning that
  the channel receiving the frame is being put on hold or retrieved from hold
- change the way channel drivers put the channel they are bridged to on hold by
  queueing a HOLD/UNHOLD control frame with the mohsuggest class as the frame
  payload instead of directly turning on moh on the bridged channel.

When a channel driver receives a HOLD frame, if the mohinterpret option is set
to "passthrough", the HOLD/UNHOLD message is passed through (currently only
supported in IAX2). Otherwise, the first choice for the moh class to use is the
class set on the channel, which will only be done with 
Set(CHANNEL(musicclass)=blah). The second choice is the control frame payload,
otherwise known as the mohsuggest setting on the channel putting you on hold.
The third choice is the mohinterpret option on the channel. Then, of course, if
none of those are set, the "default" class is used.