remove netlock, saving a lock/unlock pair on every received packet.
The race that could derive from this change is harmless and
only possible during a "sip reload" that changes the bind address,
and communication would be broken anyways so we just live with it
and the occasional error message that may occur.
(note that protection was incomplete, anyways, and did not cover
the IO handler registered on the socket).

On passing, fix a potential null pointer dereference (sipsock_read_id
could be NULL if the initial opening of the socket fails).

This is not a severe bug but probably enough for making this
change a candidate for both trunk and 1.4