Some cleanup to the CLI support routines, and an initial implementation of an enhanced CLI internal API that simplifies the implementation of the commands.
Basically, the new internal API puts the entire implementation of a CLI
command (command syntax, usage string, handler and "complete" handler)
in a single function, so it is easy to keep them all consistent.
It also passes the function a pointer to the cli entry in use,
where there is a lot of information on the entry (e.g. number of
arguments, etc) so we don't have to hardcode those in the function
itself, and changing the command string (e.g. by adding/removing
words) becomes trivial.
At this stage, the changes are fully backward compatible so
we can have both old-style and new-style commands coexisting.
The generator wrapper is not complete yet (but just a few lines of code away).

The following fixes are trunk (and possibly 1.4) candidates, 
irrespective of the above: 

- the _full_cmd string in the struct ast_cli_entry is not freed on
  unregister, thus leaking memory.
- the loop to complete the function should stop even if the selected
  entry returns NULL due to a malloc failure or end-of-choices.
- if summary == NULL, do not try to print it as %s but produce a
  <no description available> message.