a bit of cleanup and documentation of the new internal cli api, following comments from yesterday on irc/email.
This changes the module ABI because we add two fields
to the struct ast_cli_entry(), so people with binary-only
modules will need to have them rebuilt when this is committed to trunk.
Kevin says this is not a big issue.

There is a way to avoid the breakage, i am still looking at
the best way to do this before putting this stuff into trunk.