rename "peersecret" and "peermd5secret" fields in sip_pvt to "secret" and "md5secret", respectively. They don't always come from a peer so the name was misleading.
trunk candidate.

On passing: these and many other fields in the sip_pvt are
simply copied from the relevant data structure (sip_user,
sip_peer, sip_registry) when the dialog is created, and
then used once (or never).

It would make a lot more sense to either:

- just keep a reference from the sip_pvt to the source
  of this information (it is already there for sip_registry),
  so these strings can be grabbed from the source when necessary;

- if the source is dynamic, e.g. read from a database, make sure
  that the source itself has a compact representation of this info
  so we can just copy a blob of memory, and do simple lookups from there.

While this is not the worst offender in the sip_pvt size,
it does consume a significant amount of memory, and especially,
quite a bit of time (and code) to copy all the info at creation.