Merged revisions 51308-51310,51312,51314 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r51308 | rizzo | 2007-01-19 11:38:28 -0600 (Fri, 19 Jan 2007) | 3 lines

remove variable declaration in the middle of a block


................
r51309 | rizzo | 2007-01-19 11:45:46 -0600 (Fri, 19 Jan 2007) | 4 lines

sizeof() is compatible with format %d so don't be too
picky on printf formats.


................
r51310 | rizzo | 2007-01-19 11:48:48 -0600 (Fri, 19 Jan 2007) | 4 lines

in the interest of portability, avoid using %zd when all
we need is to print is an integer that fits in 16 bits.


................
r51312 | rizzo | 2007-01-19 12:00:00 -0600 (Fri, 19 Jan 2007) | 10 lines

As the comment in the diff says:

AST_INLINE_API() is a macro that takes a block of code as an argument.
Using preprocessor #directives in the argument is not supported by all
compilers, and it is a bit of an obfuscation anyways, so avoid it.
As a workaround, define a macro that produces either its argument
or nothing, and use that instead of #ifdef/#endif within the
argument to AST_INLINE_API().


................
r51314 | russell | 2007-01-19 12:06:03 -0600 (Fri, 19 Jan 2007) | 31 lines

Merged revisions 51311 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r51311 | russell | 2007-01-19 11:49:38 -0600 (Fri, 19 Jan 2007) | 23 lines

Merge the changes from the /team/group/vldtmf_fixup branch.

The main bug being addressed here is a problem introduced when two SIP
channels using SIP INFO dtmf have their media directly bridged.  So, when a
DTMF END frame comes into Asterisk from an incoming INFO message, Asterisk
would try to emulate a digit of some length by first sending a DTMF BEGIN
frame and sending a DTMF END later timed off of incoming audio.  However,
since there was no audio coming in, the DTMF_END was never generated.  This
caused DTMF based features to no longer work.

To fix this, the core now knows when a channel doesn't care about DTMF BEGIN
frames (such as a SIP channel sending INFO dtmf).  If this is the case, then
Asterisk will not emulate a digit of some length, and will instead just pass
through the single DTMF END event.

Channel drivers also now get passed the length of the digit to their digit_end
callback.  This improves SIP INFO support even further by enabling us to put
the real digit duration in the INFO message instead of a hard coded 250ms.
Also, for an incoming INFO message, the duration is read from the frame and
passed into the core instead of just getting ignored.

(issue #8597, maybe others...)

........

................