Handle station, and per-trunk on a station ring timeouts.
What is done here is that once the timeout is reached, the ringing station
will be hung up.  However, what I have not decided on, is exactly how to
handle the station after this.  When is it ok to dial the station again?
Is saying that "the station can not be called again until all trunks that
were ringing at the time of the timeout have stopped ringing" enough?

Tune in later for the exciting conclusion ...