Merged revisions 57049,57053,57055 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r57049 | murf | 2007-02-28 11:15:27 -0600 (Wed, 28 Feb 2007) | 1 line

I was surprised that I had not yet downgraded missing goto targets and macro call defs to a warning, in case they are in extensions.conf; I rectified this problem. Also, A goto in a macro to a target in a catch block was not being found; I fixed this too; the cause was that I needed to treat catch statements like an extension in the find_match code.
........
r57053 | file | 2007-02-28 11:45:50 -0600 (Wed, 28 Feb 2007) | 2 lines

Better handle timeouts when the individual speaks after everything has been played but before the timeout ends.

........
r57055 | file | 2007-02-28 11:55:03 -0600 (Wed, 28 Feb 2007) | 2 lines

Picky compiler...

........