Merged revisions 58512 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r58512 | russell | 2007-03-08 16:15:15 -0600 (Thu, 08 Mar 2007) | 5 lines

Hang up the channel that put the call on hold in the event processing thread to
avoid a race condition.  Also, if the station originated the call that it is
putting on hold, don't hang up the trunk if it was the only station on the call
and it is hanging up due to hold and not a normal hangup.

........