Merged revisions 67329 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
................
r67329 | crichter | 2007-06-05 18:11:57 +0200 (Di, 05 Jun 2007) | 9 lines

Merged revisions 67306 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.2

........
r67306 | crichter | 2007-06-05 17:39:43 +0200 (Di, 05 Jun 2007) | 1 line

simplified the EVENT_SETUP handling in the cb_events function a lot. Commented the different possibilities a bit and made functions of shared code. When the dialed extension does not exist in the extensions.conf we'll jump into the 'i' extension if this does exist, else we disconnect the call with the cause:1 = No Route to Destination.
........

................