Merged revisions 69183 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r69183 | russell | 2007-06-13 14:57:38 -0500 (Wed, 13 Jun 2007) | 9 lines

Move the logic for destroying a call when no response is received to a BYE
outside of the block that checks for FLAG_FATAL to be set.  This flag is only
set when the packet is transmitted with the reliability set to XMIT_CRITICAL
when the original packet is transmitted.  A BYE is always sent with it set
to XMIT_RELIABLE, meaning this code could never be encountered.  This resulted
in seeing some SIP channels that would never go away with the last packet
sent being a BYE.
(part of issue #9235, patch from jcmoore)

........