Merged revisions 69689 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r69689 | russell | 2007-06-18 11:15:12 -0500 (Mon, 18 Jun 2007) | 5 lines

Change the use of "echo -e" to "printf".  On systems where /bin/sh is not bash,
most of the lines in menuselect-tree were getting a "-e" at the beginning of
every line.  I'm surprised nobody noticed this, but I think the XML parser was
being very nice and ignoring them.

........