Blocked revisions 69990 via svnmerge
........
r69990 | russell | 2007-06-19 11:45:37 -0500 (Tue, 19 Jun 2007) | 12 lines

Backport fix for crashes related to subscriptions from 1.4 ...

Fix a crash that could occur when handing device state changes.
When the state of a device changes, the device state thread tells the extension
state handling code that it changed.  Then, the extension state code calls the
callback in chan_sip so that it can update subscriptions to that extension.
A pointer to a sip_pvt structure is passed to this function as the call which
needs a NOTIFY sent.  However, there was no locking done to ensure that the pvt
struct didn't disappear during this process.
(issue #9946, reported by tdonahue, patch by me, patch updated to trunk to use
 the sip_pvt lock wrappers by eliel)

........