Merged revisions 75183,75200 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r75183 | file | 2007-07-16 08:35:20 -0500 (Mon, 16 Jul 2007) | 2 lines

It is no longer required for each module that deals with a channel to call ast_module_user_hangup_all in it's unload function. The loader will automatically perform this action for it.

........
r75200 | file | 2007-07-16 09:39:29 -0500 (Mon, 16 Jul 2007) | 2 lines

Applications no longer need to call ast_module_user_add and ast_module_user_remove. This is now taken care of in the pbx_exec function outside of the application.

........