(closes issue #10209) Reported by: juggie Patches:       10209-trunk-2.patch uploaded by juggie Tested by: juggie, blitzrage
In ast_pbx_run(), mark a channel as hung up after an application returned -1, 
or when it runs out of extensions to execute.  This is so that code can detect
that this channel has been hung up for things like making sure DeadAGI is used
on actual dead channels, and is beneficial for other things, like making sure
someone doesn't try to start spying on a channel that is about to go away.