some small simplification to build_contact(), and comments on the management of the 'localnet' list (both trunk candidates.)
adapt to 64-bit ast_flags - the patch is copied from
trunk, but i am not sure it is entirely correct (both
there and here). In particulat, "flags" should be declared
as 'struct ast_flag' and not as the internal representation,
which is too error prone; and the print of 'Need Destroy'
should be just a 0/1 value not the actual flag
(this is a minor bug in trunk).