save the bits to honor the received= and rport= options in Via: headers.
This is definitely a trunk candidate, and possibly even one for 1.4
and 1.2 given that 1. the change is totally unintrusive, and
2. not honoring these options could well be considered a bug in asterisk,
and a big source of grief for users.

With this change you can pretty much forget about STUN and other
unreliable tricks for NAT traversal, at least for the SIP socket,
even in the case of a "symmetric nat", which STUN cannot handle.