Make the channel scheduler context a lot more useful than it was before. Previously, it was only used when streaming files, I think.  Now, code should be able to add callbacks to the channel scheduler and have them work under any normal circumstances.
Checking the time until the next callback has been integrated into the
ast_waitfor() core, and ast_waitfor() will return a channel if it has callbacks
that need to get executed.  ast_read() will check to see if it is time to
execute any callbacks and will do so if it is time.