two bugfixes, probably trunk candidate: - call ast_rtp_set_rtpmap_type() when found a video codec,   so we don't have to depend on the static table in rtp.c   This fixes exchanging h263+ video with x-lite.
  Probably this should be the default action, but i don't
  know enough of sdp parsing to try that.

- correctly print the audio capabilities and not the combined ones
  in one instance.

As a separate commit - there are a few places where capabilities are
still printed as decimal number, and these places should be reverted
to 0x%x format.

Once i am done with this cleanup i will commit to trunk, and
probably to 1.4 as well if we eventually want to integrate
the video support in the console driver.