Merged revisions 85316 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r85316 | russell | 2007-10-10 10:56:23 -0500 (Wed, 10 Oct 2007) | 6 lines

I introduced a new member to the ast_filestream struct in 1.4.12, but put it
in the middle of the struct, instead of at the end.  One of the Debian folks,
paravoid, pointed out that this breaks binary compatability with modules
compiled against older headers.  So, I'm moving the new member to the end
of the struct to resolve the situation.

........