Move the last instance of AST_LIBS to the only place it is used, namely main/Makefile .
I am unclear where decisions on the build environment (CFLAGS,
LDFLAGS, LIBS and so on) should be made - right now they are
split here and there.

As a first step in cleaning up this situation, i am trying to at
least collect all instances of each variable in one place.