Blocked revisions 89339 via svnmerge
........
r89339 | russell | 2007-11-16 16:26:44 -0600 (Fri, 16 Nov 2007) | 5 lines

Temporarily revert revision 89325, which added md5 magic for keeping track of
what build options were used.  We agreed that we should remove this before
making a 1.4 release, and then we can put it back in.  Then, we can take a
month or so to play around with it to get it how we want it.

........